+++ pytket/binders/circuit/main.cpp
@@ -324,10 +324,10 @@ PYBIND11_MODULE(circuit, m) {
           ":math:`P` of Pauli operations.")
       .value(
           "TermSequenceBox", OpType::TermSequenceBox,
-          "An operation defined as a set"
-          "of exponentials of the form "
-          ":math:`e^{-\\frac{i\\pi\\alpha}{2} P}` of a tensor "
-          ":math:`P` of Pauli operations.")
+          "An unordered collection of Pauli exponentials "
+          "that can be synthesised in any order, causing a "
+          "change in the unitary operation. Synthesise order "
+          "depends on the synthesis strategy chosen only.")
       .value(
           "QControlBox", OpType::QControlBox,
           "An arbitrary n-controlled operation")